backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
fastapi==0.115.6
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.5
jiter==0.12.0
langcodes==3.5.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==2.3.5
openai==2.14.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20240706
preshed==3.0.12
pycparser==2.23
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.41.3
thinc==8.3.10
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.34.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==15.0.1
wrapt==2.0.1

# source .venv/bin/activate

# npm run electron:prod
# uvicorn eurosec_ai.main:app --host 127.0.0.1 --port 48155 --reload = For running the backend
# npm run dev = to start vite
# EUROSEC_DEV_URL=http://127.0.0.1:5173 npm run electron = to run electron

# Demo in Swagger

# {
#   "user_text": "What is GDPR? Explain in 5 bullet points.",
#   "allow_cloud": true,
#   "workspace_dirs": [],
#   "preferred_files": []
# }